Move TestWatchEvents_ScopedIdentity to reside next to other WatchEvents tests
